refactor(mpi): elimina rutas en desuso
