+++ core-v2/mpi/paciente/constantes.ts
