+++ core-v2/mpi/parentesco/parentesco.routes.ts
@@ -1,7 +1,7 @@
 import { Parentesco } from './parentesco.schema';
 import { MongoQuery } from '@andes/core';
 import { ResourceBase } from '@andes/core';
-import { Auth } from '../../auth/auth.class';
+import { Auth } from '../../../auth/auth.class';
 
 class ParentescoResource extends ResourceBase {
     Model = Parentesco;
