+++ core-v2/mpi/parentesco/parentesco.schema.ts
