--- core/mpi/index.ts
@@ -1,1 +1,0 @@
-export { ParentescoRouter } from './parentesco.routes';
